Move EditAccountForm under account component
