--- client/app/account/component/index.ts
@@ -1,1 +1,0 @@
-export { AccountInfo } from './AccountInfo'
