+++ client/app/account/components/index.ts
@@ -1,1 +1,2 @@
+export { AccountInfo } from './AccountInfo'
 export { EditAccountForm } from './EditAccountForm'
