+++ client/app/account/edit/page.tsx
@@ -4,7 +4,7 @@ import React from 'react'
 import { FormContainer, Loading } from '@/_components/layout'
 import { Heading } from '@/_components/ui'
 import { useGetMeQuery } from '@/_hooks/api'
-import { EditAccountForm } from './component'
+import { EditAccountForm } from '../components'
 
 export default function Account() {
   const { data: user, isLoading } = useGetMeQuery()
