+++ client/app/account/page.tsx
@@ -5,10 +5,11 @@ import { useRouter } from 'next/navigation'
 import { FiEdit } from 'react-icons/fi'
 import { FormContainer } from '@/_components/layout'
 import { Button, Heading } from '@/_components/ui'
-import { AccountInfo } from './component'
+import { AccountInfo } from './components'
 
 export default function Account() {
   const router = useRouter()
+
   return (
     <FormContainer>
       <div className='flex flex-col gap-8 md:gap-6 '>
